Fix bug with closing popups for mobile
