+++ src/pages/home/report/ReportActionCompose/AttachmentPickerWithMenuItems.js
@@ -93,6 +93,9 @@ const propTypes = {
 
     /** Whether or not the screen is focused */
     isFocused: PropTypes.bool.isRequired,
+
+    /** A function that toggles isScrollLikelyLayoutTriggered flag for a certain period of time */
+    raiseIsScrollLikelyLayoutTriggered: PropTypes.func.isRequired,
 };
 
 const defaultProps = {
@@ -125,6 +128,7 @@ function AttachmentPickerWithMenuItems({
     onItemSelected,
     actionButtonRef,
     isFocused,
+    raiseIsScrollLikelyLayoutTriggered,
 }) {
     const theme = useTheme();
     const styles = useThemeStyles();
@@ -232,6 +236,7 @@ function AttachmentPickerWithMenuItems({
                                     <PressableWithFeedback
                                         onPress={(e) => {
                                             e.preventDefault();
+                                            raiseIsScrollLikelyLayoutTriggered();
                                             Report.setIsComposerFullSize(reportID, false);
                                         }}
                                         // Keep focus on the composer when Collapse button is clicked.
@@ -253,6 +258,7 @@ function AttachmentPickerWithMenuItems({
                                     <PressableWithFeedback
                                         onPress={(e) => {
                                             e.preventDefault();
+                                            raiseIsScrollLikelyLayoutTriggered();
                                             Report.setIsComposerFullSize(reportID, true);
                                         }}
                                         // Keep focus on the composer when Expand button is clicked.
